Chanbging types for vmx

+++ modules/core/unit/container/dsl/run.cpp
@@ -20,7 +20,7 @@ NT2_TEST_CASE( value_at )
 {
   using nt2::table;
   using nt2::of_size;
-  typedef double T;
+  typedef float T;
 
   table<T> a0(of_size(2, 3));
   table<T> a1(of_size(2, 3));
@@ -40,7 +40,7 @@ NT2_TEST_CASE( scalar_size )
 {
   using nt2::table;
   using nt2::of_size;
-  typedef double T;
+  typedef float T;
 
   table<T> a0 = T(42);
   NT2_TEST( a0.extent() == of_size(1) );
@@ -58,7 +58,7 @@ NT2_TEST_CASE( element_wise )
 {
   using nt2::table;
   using nt2::of_size;
-  typedef double T;
+  typedef float T;
 
   table<T> a0;
   table<boost::dispatch::meta::as_integer<T>::type> a1;
